requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-unfold==0.38.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
